requirements.txt
django<2
django-import-export
pillow
docutils
djangorestframework
markdown
